Move product form manager logic to FormManager class
